fix animation

+++ examples/ios/TrainControl/TrainControl/view/WMGaugeView.h
@@ -122,9 +122,9 @@ IB_DESIGNABLE
 /**
  * WMGaugeView public functions
  */
-- (void)setValue:(float)value animated:(BOOL)animated;
-- (void)setValue:(float)value animated:(BOOL)animated completion:(void (^)(BOOL finished))completion;
-- (void)setValue:(float)value animated:(BOOL)animated duration:(NSTimeInterval)duration;
-- (void)setValue:(float)value animated:(BOOL)animated duration:(NSTimeInterval)duration completion:(void (^)(BOOL finished))completion;
+- (void)setValue:(CGFloat)value animated:(BOOL)animated;
+- (void)setValue:(CGFloat)value animated:(BOOL)animated completion:(void (^)(BOOL finished))completion;
+- (void)setValue:(CGFloat)value animated:(BOOL)animated duration:(NSTimeInterval)duration;
+- (void)setValue:(CGFloat)value animated:(BOOL)animated duration:(NSTimeInterval)duration completion:(void (^)(BOOL finished))completion;
 
 @end
